+++ students/FridaKvale/SvadaTekst/tekster.txt
